Renaming the Real implementation of the server to HttpServer, and remove the initial interface put in place
